updated fabric api, moved a bunch of stuff to item module
