+++ felt-item-api/build.gradle
@@ -1,2 +1,6 @@
 archivesBaseName = "felt-item-api"
 version = getSubprojectVersion(project)
+
+dependencies {
+    implementation(project(":felt-tool-api"))
+}
