+++ felt-item-api/src/main/java/io/github/feltmc/feltapi/api/item/extensions/ArmorExtension.java
@@ -1,13 +1,13 @@
-package io.github.feltmc.feltapi.api.tool.extensions;
+package io.github.feltmc.feltapi.api.item.extensions;
 
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.item.ItemStack;
 
-import java.sql.Ref;
-
-public interface ArmorTextureExtension {
+public interface ArmorExtension {
     default String getArmorTexture(ItemStack stack, Entity entity, EquipmentSlot slot, String type) {
         return null;
     }
+
+
 }
