+++ felt-item-api/src/main/java/io/github/feltmc/feltapi/mixin/item/ItemStackMixin.java
@@ -1,4 +1,4 @@
-package io.github.feltmc.feltapi.mixin.tool.common;
+package io.github.feltmc.feltapi.mixin.item;
 
 import io.github.feltmc.feltapi.api.tool.extensions.DamageableItemExtension;
 import net.minecraft.item.Item;
