+++ felt-item-api/src/main/resources/itemapi.mixins.json
@@ -4,7 +4,16 @@
   "package": "io.github.feltmc.feltapi.mixin.item",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-    "PacketByteBufMixin"
+    "PacketByteBufMixin",
+    "EnchantmentHelperMixin",
+    "EnchantmentMixin",
+    "ItemStackMixin",
+    "PlayerEntityMixin",
+    "ServerPlayerInteractionManagerMixin"
+  ],
+  "client": [
+    "client.ArmorFeatureRendererMixin",
+    "client.ClientPlayerInteractionManagerMixin"
   ],
   "injectors": {
     "defaultRequire": 1
