+++ felt-tool-api/src/main/resources/toolapi.mixins.json
@@ -4,19 +4,9 @@
   "package": "io.github.feltmc.feltapi.mixin.tool",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-    "common.AxeMixin",
-    "common.EnchantmentHelperMixin",
-    "common.EnchantmentMixin",
-    "common.HoeMixin",
-    "common.ItemStackMixin",
-    "io.github.feltmc.feltapi.mixin.item.PacketByteBufMixin",
-    "common.PlayerEntityMixin",
-    "common.ServerPlayerInteractionManagerMixin",
-    "common.ShovelMixin"
-  ],
-  "client": [
-    "client.ArmorFeatureRendererMixin",
-    "client.ClientPlayerInteractionManagerMixin"
+    "AxeMixin",
+    "HoeMixin",
+    "ShovelMixin"
   ],
   "injectors": {
     "defaultRequire": 1
