+++ gradle.properties
@@ -19,5 +19,5 @@ org.gradle.jvmargs=-Xmx1G
 	felt-shared-lifecycle-events-version=1.0.0
 
 # Dependencies
-	fabric_version=0.47.8+1.18.2
+	fabric_version=0.55.1+1.18.2
 	modmenu_version=3.1.0
